+++ packages/components/organisms/f-content-cards/CHANGELOG.md
@@ -3,6 +3,17 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v6.0.0-beta.1
+------------------------------
+*September 20, 2021*
+
+### Changed
+- Updating braze adapter version
+- Adding finer detailed logs
+- Updated unit tests
+
+
 v6.0.0-beta.0
 ------------------------------
 *September 20, 2021*